Add configuration , rename
