+++ notify.conf
@@ -1,0 +1,6 @@
+{
+    "notifierId" : "287954569844097024",
+    "notificationChannelId" : "1041794511929618462",
+    "notificationEmoji" : "✅"
+
+}
